refactor: 장바구니관련 통신 라이브러리 Retrofit으로 변경

+++ app/src/main/java/woowacourse/shopping/data/recentproduct/RecentProductDao.kt
@@ -149,7 +149,7 @@ class RecentProductDao(
                 getLong(getColumnIndexOrThrow(RecentProductContract.TABLE_COLUMN_VIEWED_DATE_TIME))
 
             return RecentProduct(
-                productId = productId,
+                productId = productId.toLong(),
                 productImageUrl = productImageUrl,
                 productName = productName,
                 productPrice = productPrice,
