+++ domain/src/main/java/com/example/domain/CartProduct.kt
@@ -1,8 +1,8 @@
 package com.example.domain
 
 class CartProduct(
-    val id: Int,
-    val productId: Int,
+    val id: Long,
+    val productId: Long,
     val productImageUrl: String,
     val productName: String,
     val productPrice: Int,
